+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Allow to reuse same UDP address
   * Socket events on Bus instance
+  * Enable UDP broadcast
 
 ## 0.5.0